requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-limiter==3.5.0
python-dotenv==1.0.1
Werkzeug==3.0.1
razorpay==1.4.2
bleach==6.1.0
qrcode==7.4.2

# Production server
gunicorn==22.0.0
SQLAlchemy==2.0.29
Flask-Migrate==4.0.7
alembic==1.13.1
